-Splits repository interfaces for better modularity
