+++ src/datastore/interfaces/repository/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Copyright - 2025 - Maleesha Gimshan (github.com/maleeshagimshan98)
+ */
+
+export { UsersRepositroy } from './users.repository';
+export { ConversationsRepository } from './conversations.repository';
+export { MessagesRepository } from './messages.repository';
